requirements.txt
crewai>=0.28.0
streamlit>=1.31.1
chromadb>=0.4.20
chroma-hnswlib>=0.7.3
sentence-transformers>=2.2.2
groq>=0.4.0
python-decouple>=3.8
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.8.0
chardet>=5.2.0
